+++ foundation_headless/init.lua
@@ -1,3 +1,15 @@
-local mod = foundation.new_module("foundation_headless", "1.2.1")
+--
+-- Foundation Headless
+--
+local mod = foundation.new_module("foundation_headless", "2.0.0")
 
-mod:require("fake_meta_ref.lua")
+--- @namespace foundation.com.headless
+foundation.com.headless = foundation.com.headless or {}
+
+mod:require("meta_data_ref.lua")
+mod:require("item_stack.lua")
+mod:require("inv_ref.lua")
+
+if foundation.com.Luna then
+  mod:require("tests.lua")
+end