+++ foundation_headless/tests.lua
@@ -1,0 +1,3 @@
+local mod = assert(foundation_headless)
+
+mod:require("tests/item_stack_test.lua")