+++ foundation_mt/tests.lua
@@ -1,0 +1,5 @@
+local mod = assert(foundation_mt)
+
+mod:require("tests/metaref_test.lua")
+mod:require("tests/item_stack_test.lua")
+mod:require("tests/inventory_list_test.lua")